avoid infinite blocking in TcpConnectionSpec, #21375
* problem occurs when network is offline
* use socket timeout
* additional cleanup of socket utils
