--- akka-docs/rst/scala/code/docs/utils/TestUtils.scala
@@ -1,24 +1,0 @@
-/**
- * Copyright (C) 2009-2016 Lightbend Inc. <http://www.lightbend.com>
- */
-
-package docs.utils
-
-import java.net.InetSocketAddress
-import java.nio.channels.ServerSocketChannel
-
-object TestUtils {
-  def temporaryServerAddress(interface: String = "127.0.0.1"): InetSocketAddress = {
-    val serverSocket = ServerSocketChannel.open()
-    try {
-      serverSocket.socket.bind(new InetSocketAddress(interface, 0))
-      val port = serverSocket.socket.getLocalPort
-      new InetSocketAddress(interface, port)
-    } finally serverSocket.close()
-  }
-
-  def temporaryServerHostnameAndPort(interface: String = "127.0.0.1"): (String, Int) = {
-    val socketAddress = temporaryServerAddress(interface)
-    socketAddress.getHostName -> socketAddress.getPort // TODO getHostString in Java7
-  }
-}
